backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.32
alembic==1.13.2
python-multipart==0.0.9
bcrypt==4.2.0
PyJWT==2.9.0
httpx==0.27.2
python-dotenv==1.0.1
